Final commit
Fixed a bug with the X buttons when closing the windows it still kept the cursors open when it was another window opened in some instances, in others were closing the cursor without being necessary. Made the Skill Tree buttons work and now adds the corresponding stats to the status page. Buttons work on tier and the previous tier needs to be achieved to progress.